Divided AddLayerPopOver into subcomponents(container/component structure) for better code management. And a bit of code cleanup.
