+++ src/components/Sidebar.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import AddLayerAndSearch from './AddLayerAndSearch';
-import AddLayer from './AddLayer'
+import AddLayerPopOver from './AddLayerPopOver'
 
 class Sidebar extends Component
 {
@@ -27,7 +27,7 @@ class Sidebar extends Component
                 <AddLayerAndSearch showAddLayerPopOver={ this.showAddLayerPopOver } />
                 {
                     this.state.addLayerPopOverIsShown && 
-                    <AddLayer hideAddLayerPopOver={ this.hideAddLayerPopOver } />
+                    <AddLayerPopOver hideAddLayerPopOver={ this.hideAddLayerPopOver } />
                 }
                 
             </div>
